+++ buession-redis/src/main/java/com/buession/redis/client/connection/lettuce/LettuceClusterConnection.java
@@ -576,7 +576,7 @@ public class LettuceClusterConnection extends AbstractLettuceRedisConnection imp
 	}
 
 	@Override
-	public boolean isConnect() {
+	public boolean isConnected() {
 		return delegate != null && delegate.isOpen();
 	}